separate express app and server logic
