+++ backend/src/api/app.ts
@@ -1,0 +1,22 @@
+import express, { Request, Response } from 'express';
+import morgan from 'morgan';
+import helmet from 'helmet';
+import cors from 'cors';
+import 'dotenv/config';
+
+const app = express();
+
+/**
+ * Dependencies configurations
+ */
+
+app.use(morgan('dev'));
+app.use(helmet());
+app.use(cors());
+app.use(express.json());
+
+app.get('/', (req: Request, res: Response) => {
+    res.status(200).json({ message: 'Hello, world!' });
+});
+
+export default app;
